+++ modules/edit-modes/src/index.js
@@ -1,11 +1,9 @@
 // @flow
 
-export { EditMode, BaseEditMode } from './lib/edit-mode.js';
+export { EditMode } from './lib/edit-mode.js';
 export { ViewMode } from './lib/view-mode.js';
 export { DrawPolygonMode } from './lib/draw-polygon-mode.js';
 export { ImmutableFeatureCollection } from './lib/immutable-feature-collection.js';
-
-export type { ModeState } from './lib/edit-mode.js';
 
 export type {
   ScreenCoordinates,
@@ -14,7 +12,8 @@ export type {
   ClickEvent,
   PointerMoveEvent,
   StartDraggingEvent,
-  StopDraggingEvent
+  StopDraggingEvent,
+  ModeProps
 } from './types.js';
 
 export type {