+++ modules/edit-modes/src/types.js
@@ -63,3 +63,26 @@ export type PointerMoveEvent = {
   pointerDownMapCoords: ?Position,
   sourceEvent: any
 };
+
+export type ModeProps<TData> = {
+  // The data being edited, this can be an array or an object
+  data: TData,
+
+  // Additional configuration for this mode
+  modeConfig: any,
+
+  // The indexes of the selected features
+  selectedIndexes: number[],
+
+  // The cursor type, as a [CSS Cursor](https://developer.mozilla.org/en-US/docs/Web/CSS/cursor)
+  cursor: ?string,
+
+  // The last pointer move event that occurred
+  lastPointerMoveEvent: PointerMoveEvent,
+
+  // Callback used to notify applications of an edit action
+  onEdit: (editAction: EditAction<TData>) => void,
+
+  // Callback used to update cursor
+  onUpdateCursor: (cursor: ?string) => void
+};